fix: apply server properties to handshake
