+++ internal/core/amqp/basic.go
@@ -44,7 +44,7 @@ type BasicProperties struct {
 	Reserved        string         // shortstr
 }
 
-func (props *BasicProperties) EncodeBasicProperties() ([]byte, uint16, error) {
+func (props *BasicProperties) encodeBasicProperties() ([]byte, uint16, error) {
 	var buf bytes.Buffer
 	var flags uint16
 
@@ -137,6 +137,28 @@ func (props *BasicProperties) EncodeBasicProperties() ([]byte, uint16, error) {
 		}
 	}
 	return buf.Bytes(), flags, nil
+}
+
+func decodeBasicGetFlags(octet byte) map[string]bool {
+	flags := make(map[string]bool)
+	flagNames := []string{"noAck", "flag2", "flag3", "flag4", "flag5", "flag6", "flag7", "flag8"}
+
+	for i := 0; i < 8; i++ {
+		flags[flagNames[i]] = (octet & (1 << uint(7-i))) != 0
+	}
+
+	return flags
+}
+
+func decodeBasicPublishFlags(octet byte) map[string]bool {
+	flags := make(map[string]bool)
+	flagNames := []string{"mandatory", "immediate", "flag3", "flag4", "flag5", "flag6", "flag7", "flag8"}
+
+	for i := 0; i < 8; i++ {
+		flags[flagNames[i]] = (octet & (1 << uint(7-i))) != 0
+	}
+
+	return flags
 }
 
 func encodeShortStr(buf *bytes.Buffer, value string) error {
